refacto(llm): CLS image processing files renaming
