+++ apps/ledger-live-mobile/src/components/CustomImage/injectedCode/imageFromDeviceProcessing.ts
