+++ apps/ledger-live-mobile/src/components/RootNavigator/types/CustomImageNavigator.ts
@@ -2,7 +2,10 @@ import { Device } from "@ledgerhq/types-devices";
 import { type CLSSupportedDeviceModelId } from "@ledgerhq/live-common/device/use-cases/isCustomLockScreenSupported";
 import { ScreenName } from "~/const";
 import { CropResult } from "../../CustomImage/ImageCropper";
-import { ProcessorPreviewResult, ProcessorRawResult } from "../../CustomImage/ImageProcessor";
+import {
+  ProcessorPreviewResult,
+  ProcessorRawResult,
+} from "../../CustomImage/ImageToDeviceProcessor";
 import { GalleryNFT, ImageFileUri, ImageUrl, ImageType } from "../../CustomImage/types";
 
 type BaseParams = {
